requirements.txt
Django==1.9.1
djangorestframework==3.3.2
dj-database-url==0.3.0
psycopg2==2.6.1
raven==5.10.0
gunicorn==19.4.5
django-filter==0.12.0
dj-static==0.0.6
celery==3.1.19
django-celery==3.1.17
redis==2.10.5
pytz==2015.7
